Promote gametypes in XD_MAP, XD_MAPQUEUE and XD_REQMAPQUEUE to UINT16
Since new-voting repermitted custom gametypes to go beyond the bounds of a UINT8, here's some extra anti-footgun protection.

+++ src/g_game.h
@@ -58,7 +58,7 @@ typedef enum
 struct roundentry_t
 {
 	UINT16 mapnum;				// Map number at this position
-	UINT8 gametype;				// Gametype we want to play this in
+	UINT16 gametype;			// Gametype we want to play this in
 	boolean encore;				// Whether this will be flipped
 	boolean rankrestricted;		// For grand prix progression
 };
